frame buffer 8 bit color
color not working

+++ mycode/wireframe_cube_opt/lib/FrameBuffer/FrameBuffer.cpp
@@ -2,7 +2,7 @@
 
 FrameBuffer::FrameBuffer(int16_t w, int16_t h) : Elegoo_GFX(w, h) {
   // (viene chiamato prima il costruttore di elegoo_GFX e poi questo)
-  bufferPointer = new uint16_t[_width * _height]; // Allocazione del buffer array
+  bufferPointer = new uint8_t[_width * _height]; // Allocazione del buffer array
   resetBuffer();
 }
 
@@ -20,7 +20,7 @@ void FrameBuffer::drawBuffer(Elegoo_TFTLCD *tft) {
   for (int16_t y = 0; y < _height; y++) {
     for (int16_t x = 0; x < _width; x++) {
       uint32_t index = y * _width + x;
-      uint16_t color = bufferPointer[index];
+      uint8_t color = bufferPointer[index];
       if (color == 0) {
         color = 0x0000; // default black
       }
@@ -30,7 +30,7 @@ void FrameBuffer::drawBuffer(Elegoo_TFTLCD *tft) {
 }
 
 void FrameBuffer::resetBuffer() {
-  memset(bufferPointer, 0x0000, _width * _height * sizeof(uint16_t));
+  memset(bufferPointer, 0x0000, _width * _height * sizeof(uint8_t));
 }
 
 char* FrameBuffer::hello() {
